refactor(dashboard): 집계 계산을 metrics 모듈로 분리
화면 숫자를 정하는 규칙(미분류를 분모에서 빼는 것, 칸 경계를 내림하는 것)이
렌더 코드 사이에 흩어져 있어서 한쪽만 바뀔 여지가 있었습니다. 두 감정 차트가
공유하는 components/feedback/sentiment.ts와 같은 방식으로 뺐습니다.

vitest가 environment: 'node'라 컴포넌트 안에서는 테스트를 붙일 수 없었는데,
순수 함수가 되면서 테스트가 가능해집니다.

감정별 개수와 긍정 비율은 summarizeSentiments 하나로 묶었습니다. 세는 일과
나누는 일이 떨어져 있으면 분모를 다시 만들 때 미분류를 빼는 규칙이
따라오지 않습니다.

Refs #115
